pyproject.toml
[project]
name = "docintelligence"
version = "0.1.0"
description = ""
authors = [
    {name = "Jeff Liu",email = "jeff.liu@outlook.co.nz"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai (>=1.75.0,<2.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "streamlit (>=1.44.1,<2.0.0)",
    "snowflake-connector-python (>=3.14.1,<4.0.0)",
    "pyarrow (<19.0.0)"
]

[tool.poetry]
packages = [{include = "docintelligence", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
